extract array helpers
